requirements.txt
fastapi
uvicorn
pydantic
openai
python-jose[cryptography]
passlib[bcrypt]
python-multipart